+++ ceammc/ext/src/base/prop_set.h
@@ -1,5 +1,5 @@
 /*****************************************************************************
- * Copyright 2017 Serge Poltavsky. All rights reserved.
+ * Copyright 2020 Serge Poltavsky. All rights reserved.
  *
  * This file may be distributed under the terms of GNU Public License version
  * 3 (GPL v3) as defined by the Free Software Foundation (FSF). A copy of the
@@ -11,35 +11,23 @@
  * contact the author of this file, or the owner of the project in which
  * this file belongs to.
  *****************************************************************************/
-#ifndef FLOW_COUNT_H
-#define FLOW_COUNT_H
+#ifndef PROP_SET_H
+#define PROP_SET_H
 
 #include "ceammc_object.h"
-
 using namespace ceammc;
 
-class FlowCount : public BaseObject {
-    IntProperty* counter_;
-    int default_value_;
+class PropSet : public BaseObject {
+    std::vector<t_symbol*> props_;
 
 public:
-    FlowCount(const PdArgs& a);
+    PropSet(const PdArgs& args);
 
-    void onBang() override;
-    void onFloat(t_float f) override;
-    void onSymbol(t_symbol* s) override;
-    void onList(const AtomList& l) override;
-    void onAny(t_symbol* s, const AtomListView& l) override;
-    void onData(const Atom&) override;
-    void onInlet(size_t, const AtomList&) override;
-
-    bool processAnyProps(t_symbol* s, const AtomListView&) override;
-    void initDone() override;
-
-private:
-    void tick();
+    void parseProperties() override;
+    void onInlet(size_t n, const AtomList& lst) override;
+    const char* annotateInlet(size_t n) const override;
 };
 
-void setup_flow_count();
+void setup_prop_set();
 
-#endif // FLOW_COUNT_H
+#endif // PROP_SET_H