+++ ceammc/ext/src/flow/flow_tee_tilde.cpp
@@ -61,9 +61,11 @@ bool FlowTeeTilde::processAnyProps(t_symbol*, const AtomListView&)
     return false;
 }
 
-extern "C" void setup_flow0x2etee_tilde()
+void setup_flow_tee_tilde()
 {
     SoundExternalFactory<FlowTeeTilde> obj("flow.tee~");
     obj.addAlias("tee~");
     obj.addAlias("~->");
+
+    obj.setXletsInfo({ "intermixed control/signal" }, { "output signal", "output control messages" });
 }