+++ ceammc/ext/src/flow/mod_flow.cpp
@@ -31,14 +31,14 @@
 #include "flow_tee_tilde.h"
 
 void setup_flow_match();
+void setup_flow_delay();
 
 void ceammc_flow_setup()
 {
-    setup_flow0x2etee_tilde();
-
     setup_flow_append();
     setup_flow_change();
     setup_flow_count();
+    setup_flow_delay();
     setup_flow_demultiplex();
     setup_flow_demultiplex2_tilde();
     setup_flow_demultiplex_tilde();
@@ -65,4 +65,5 @@ void ceammc_flow_setup()
     setup_flow_split();
     setup_flow_sync();
     setup_flow_sync_pack();
+    setup_flow_tee_tilde();
 }